ignore repeat IR commands

+++ src/Control Tasks/IRControlTask.hpp
@@ -1,6 +1,8 @@
 #ifndef IR_CONTROL_TASK_HPP
 #define IR_CONTROL_TASK_HPP
 
+#include <stdint.h>
+
 class IRControlTask {
 public:
     void begin();
@@ -10,7 +12,7 @@ private:
     void parse_command();
     void handle_overflow();
 
-    int button_selected = -1;
+    uint16_t button_selected = -1;
 };
 
 #endif
